payments by card and cash

+++ tgbot/handlers/users/product_order.py
@@ -2,12 +2,15 @@ from typing import Union
 from aiogram import types
 from aiogram.dispatcher import FSMContext
 from decimal import Decimal
+
+from tgbot.handlers.users.payments import show_invoice
 from tgbot.keyboards.default.menu_kb import menu
-from tgbot.keyboards.inline.callback_datas import user_address_callback
-from tgbot.keyboards.inline.order_keyboards import generate_addresses_keyboard, gen_check_keyboard
+from tgbot.keyboards.inline.callback_datas import user_address_callback, shipping_callback, payment_callback
+from tgbot.keyboards.inline.order_keyboards import generate_addresses_keyboard, gen_check_keyboard, \
+    gen_shipping_keyboard, gen_payment_keyboard
 from tgbot.loader import dp, bot
 from tgbot.states.order_states import OrderStates
-from tgbot.utils.cart_product_utils import create_cart_list, gen_total_price, wipe_cart_data
+from tgbot.utils.cart_product_utils import create_cart_list, gen_total_price, wipe_state_data
 from tgbot.utils.db_api.quick_commands import get_user, get_ordered_products
 from tgbot.utils.db_api.schemas.db_tables import OrdersGino, OrderProductGino, TgUserGino, UserAddresses
 from tgbot.utils.generate_order_number import generate_order_number
@@ -61,10 +64,32 @@ async def accept_address(message: Union[types.Message, types.CallbackQuery], sta
         await state.update_data(user_address=address)
         await state.reset_state(with_data=False)
         await message.answer()
+    # answer = "Почти готово! Поделитесь с нами вашим номером телефона, " \
+    #          "чтобы мы могли связаться с вами. Или напишите его ниже в формате +998**9999999"
+    await bot.send_message(chat_id=chat_id, text="Чем Доставлять?", reply_markup=gen_shipping_keyboard())
+    await OrderStates.Shipping.set()
+
+
+@dp.callback_query_handler(shipping_callback.filter(), state=OrderStates.Shipping)
+async def shipping_method(call: types.CallbackQuery, callback_data: dict, state: FSMContext):
+    shipping = callback_data.get("name")
+    if shipping == "courier":
+        await state.update_data(courier=True)
+    await bot.send_message(call.message.chat.id, text='Как удобно оплатить?', reply_markup=gen_payment_keyboard())
+    await OrderStates.Payment.set()
+    await call.answer()
+
+
+@dp.callback_query_handler(payment_callback.filter(), state=OrderStates.Payment)
+async def payment_method(call: types.CallbackQuery, callback_data: dict, state: FSMContext):
+    payment = callback_data.get("name")
+    if payment == "card":
+        await state.update_data(pay_by_card=True)
     answer = "Почти готово! Поделитесь с нами вашим номером телефона, " \
              "чтобы мы могли связаться с вами. Или напишите его ниже в формате +998**9999999"
-    await bot.send_message(chat_id=chat_id, text=answer)
+    await bot.send_message(call.message.chat.id, text=answer)
     await OrderStates.Phone_Number.set()
+    await call.answer()
 
 
 @dp.message_handler(state=OrderStates.Phone_Number)
@@ -79,8 +104,8 @@ async def is_correct(chat_id: str, **kwargs):
     async with state.proxy() as state_data:
         address = state_data["user_address"]
         phone_number = state_data['phone_number']
-    cart_list = await create_cart_list(state)
-    answer = "Всё верно?\n\n" + f"{cart_list}\n\nАдрес доставки: {address}\n\nКонтакт: {phone_number}"
+        cart_list = await create_cart_list(state)
+        answer = "Всё верно?\n\n" + f"{cart_list}\n\nАдрес доставки: {address}\n\nКонтакт: {phone_number}"
     await bot.send_message(chat_id=chat_id, text=answer, reply_markup=gen_check_keyboard())
     await state.reset_state(with_data=False)
 
@@ -88,21 +113,26 @@ async def is_correct(chat_id: str, **kwargs):
 @dp.callback_query_handler(text="make_order")
 async def make_order(call: types.CallbackQuery, state: FSMContext):
     chat_id = call.from_user.id
+    state_data = await state.get_data()
     await create_order_db(state=state)
     await ordered_product_list(state=state)
+    if state_data.get("pay_by_card"):
+        await show_invoice(chat_id, state)
+        await call.answer()
+        return
     state_data = await state.get_data()
     order_number = state_data.get("order_number")
     answer = f"Спасибо, номер заказа {order_number}! Наш менеджер свяжется с вами для уточнения всех деталей."
     await bot.send_message(chat_id=chat_id, text=answer)
     await order_notify(state)
-    await wipe_cart_data(state, products=True)
+    await wipe_state_data(state, products=True)
     await bot.delete_message(call.message.chat.id, call.message.message_id)
     await call.answer()
 
 
 @dp.callback_query_handler(text="cancel_order")
 async def cancel_order(call: types.CallbackQuery, state: FSMContext):
-    await wipe_cart_data(state)
+    await wipe_state_data(state)
     await bot.send_message(call.message.chat.id, text='Мы перебросили вас в главное меню, но ваша корзина на месте',
                            reply_markup=menu)
     await bot.delete_message(call.message.chat.id, call.message.message_id)
